blur: Interpolate tap UV coordinates in vertex shader
Instead of calculating the UV coordinate to sample for each blur tap in
the vertex shader, calculate them as varying attributes in the vertex
shader and let the GPU interpolate them for each fragment. This allows
the GPU to prefetch pixel data because we no longer rely on dependent
texture reads.

When tested with 2 layers of another 6-pass blur implementation, this
saves ~100 µs of rendering time on an Adreno 640 GPU at 1440x3040
resolution.

Change-Id: Id2a700decef060ea23d1d5e2cccaea864581bf16
Signed-off-by: Mohammad Hasan Keramat J <ikeramat@protonmail.com>
Signed-off-by: Dmitrii <bankersenator@gmail.com>
Signed-off-by: HeroBuxx <me@herobuxx.me>